requirements.txt
mysql-connector-python
clickhouse-driver
psycopg2-binary
duckdb